requirements.txt
absl-py==2.3.1
actionlib==1.14.3
angles==1.9.14
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.11.0
apturl==0.5.2
asttokens==3.0.0
astunparse==1.6.3
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.8.2
bondpy==1.8.7
Brlapi==0.7.0
cachetools==5.5.2
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
constantly==15.1.0
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
defusedxml==0.6.0
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
docutils==0.16
duplicity==0.8.12.0
dynamic-reconfigure==1.7.6
empy==3.3.2
entrypoints==0.3
exceptiongroup==1.3.0
executing==2.2.0
fail2ban==0.11.1
fastapi==0.118.0
fasteners==0.14.1
flask==3.0.3
flask-sqlalchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.57.0
future==0.18.2
gast==0.4.0
gazebo-plugins==2.9.3
gazebo-ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
google-api-core==2.25.2
google-auth==2.41.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.16.0
h5py==3.11.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==16.10.1
interactive-markers==1.12.2
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
jupyter-client==8.6.3
jupyter-core==5.8.1
keras==2.13.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.8
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
lockfile==0.12.2
louis==3.12.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
message-filters==1.17.4
monotonic==1.5
mpi4py==3.0.3
nest-asyncio==1.6.0
netifaces==0.10.4
nose==1.3.7
numpy==1.24.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.4.0
packaging==25.0
pandas==2.0.3
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.3
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydantic==2.10.6
pydantic-core==2.27.2
pydot==1.4.1
pygments==2.19.2
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.6
python-snappy==0.5.3
PyTrie==0.2
pytz==2025.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==27.0.0
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resource-retriever==1.12.10
roman==2.0.0
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosdep-modules==0.25.1
rosdistro-modules==1.0.1
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rsa==4.9.1
rviz==1.14.26
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
sensor-msgs==1.13.2
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
sniffio==1.3.1
sqlalchemy==2.0.48
ssh-import-id==5.10
stack-data==0.6.3
starlette==0.44.0
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
threadpoolctl==3.5.0
topic-tools==1.17.4
tornado==6.4.2
traitlets==5.14.3
Twisted==18.9.0
txaio==2.10.0
typing-extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
u-msgpack-python==2.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.33.0
wadllib==1.3.3
wcwidth==0.2.13
werkzeug==3.0.6
wrapt==1.17.3
wsaccel==0.6.2
xacro==1.14.20
xkit==0.0.0
zipp==3.20.2
zope.interface==4.7.1
